requirements.txt
gradio==4.7.1
openai==1.26.0
python-dotenv==1.0.0
loguru==0.7.2